fix(remix): Always instrument callRouteAction in channel integration
`instrumentRemix` only subscribed to `callRouteAction` when
`actionFormDataAttributes` was configured, so ACTION spans were never created
unless form-data capture was opted into. The OTel instrumentation always patched
`callRouteAction` and used those keys only for optional attribute extraction.

Always subscribe; the request clone + async form-data read now happen only when
capture is configured. Adds a regression test (separate file so the channel
subscribes with no form-data config) asserting ACTION spans are still created.
